+++ spec/requests/api/stores_controller_spec.rb
@@ -30,7 +30,7 @@ describe Api::StoresController, swagger_doc: "api.yaml" do
         context "with success response from the proxied server", vcr: "api/stores_post_success" do
           context "when proxy mode is active by default" do # rubocop:disable RSpec/EmptyExampleGroup
             run_test! do
-              expect(response.body).to eq('["origin_response_body"]')
+              expect(response.body).to eq('["origin_result"]')
             end
           end
 
@@ -38,7 +38,7 @@ describe Api::StoresController, swagger_doc: "api.yaml" do
             include_context "with flipper enabled", "api/stores#index:mirror"
 
             run_test! do
-              expect(response.body).to eq('["origin_response_body"]')
+              expect(response.body).to eq('["origin_result"]')
             end
           end
         end